Move validation-related things to common
As this seems a more logical place
