+++ src/professions/admin/add-profession/legislation.controller.ts
@@ -15,8 +15,8 @@ import { Permissions } from '../../../common/permissions.decorator';
 import { Validator } from '../../../helpers/validator';
 import { Legislation } from '../../../legislations/legislation.entity';
 import { UserPermission } from '../../../users/user.entity';
-import { ValidationExceptionFilter } from '../../../validation/validation-exception.filter';
-import { ValidationFailedError } from '../../../validation/validation-failed.error';
+import { ValidationExceptionFilter } from '../../../common/validation/validation-exception.filter';
+import { ValidationFailedError } from '../../../common/validation/validation-failed.error';
 import { ProfessionsService } from '../../professions.service';
 import LegislationDto from './dto/legislation.dto';
 import { LegislationTemplate } from './interfaces/legislation.template';
