+++ src/professions/admin/add-profession/qualification-information.controller.ts
@@ -15,7 +15,7 @@ import { Permissions } from '../../../common/permissions.decorator';
 import { Validator } from '../../../helpers/validator';
 import { Qualification } from '../../../qualifications/qualification.entity';
 import { UserPermission } from '../../../users/user.entity';
-import { ValidationFailedError } from '../../../validation/validation-failed.error';
+import { ValidationFailedError } from '../../../common/validation/validation-failed.error';
 import { ProfessionsService } from '../../professions.service';
 import { MethodToObtainQualificationRadioButtonsPresenter } from '../method-to-obtain-qualification-radio-buttons.presenter';
 import { YesNoRadioButtonArgsPresenter } from '../yes-no-radio-buttons-presenter';
