+++ src/professions/admin/add-profession/regulated-activities.controller.ts
@@ -11,7 +11,7 @@ import {
 import { Response } from 'express';
 import { AuthenticationGuard } from '../../../common/authentication.guard';
 import { Validator } from '../../../helpers/validator';
-import { ValidationFailedError } from '../../../validation/validation-failed.error';
+import { ValidationFailedError } from '../../../common/validation/validation-failed.error';
 import { Profession } from '../../profession.entity';
 import { ProfessionsService } from '../../professions.service';
 import { RegulatedActivitiesDto } from './dto/regulated-activities.dto';
